% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ByEvent.R
\name{ByEvent}
\alias{ByEvent}
\title{Cut by event}
\description{
A class for event-based cut
}
\section{Super class}{
\code{\link[gsdcut:Cut]{gsdcut::Cut}} -> \code{ByEvent}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{event}}{number of event to cut the interim/final analysis.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ByEvent-validate}{\code{ByEvent$validate()}}
\item \href{#method-ByEvent-new}{\code{ByEvent$new()}}
\item \href{#method-ByEvent-summary}{\code{ByEvent$summary()}}
\item \href{#method-ByEvent-clone}{\code{ByEvent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByEvent-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ByEvent-validate}{}}}
\subsection{Method \code{validate()}}{
Validate input parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByEvent$validate(event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{number of event to cut the interim/final analysis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByEvent-new"></a>}}
\if{latex}{\out{\hypertarget{method-ByEvent-new}{}}}
\subsection{Method \code{new()}}{
TBA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByEvent$new(event)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{number of event to cut the interim/final analysis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Exponential} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByEvent-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ByEvent-summary}{}}}
\subsection{Method \code{summary()}}{
TBA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByEvent$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByEvent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ByEvent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByEvent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
