% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ByHybrid.R
\name{ByHybrid}
\alias{ByHybrid}
\title{Cut by time}
\description{
A class for event-based cut
}
\section{Super class}{
\code{\link[gsdcut:Cut]{gsdcut::Cut}} -> \code{ByHybrid}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time}}{calendar time to cut the interim/final analysis.}

\item{\code{event}}{target events to cut the interim/final analysis.}

\item{\code{planned_total_duration}}{planned duration cutoff}

\item{\code{min_followup}}{minimum follow-up duration}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ByHybrid-clone}{\code{ByHybrid$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="gsdcut" data-topic="Cut" data-id="summary"><a href='../../gsdcut/html/Cut.html#method-Cut-summary'><code>gsdcut::Cut$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gsdcut" data-topic="Cut" data-id="validate"><a href='../../gsdcut/html/Cut.html#method-Cut-validate'><code>gsdcut::Cut$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByHybrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ByHybrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByHybrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
