% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ByTime.R
\name{ByTime}
\alias{ByTime}
\title{Cut by time}
\description{
A class for event-based cut
}
\section{Super class}{
\code{\link[gsdcut:Cut]{gsdcut::Cut}} -> \code{ByTime}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time}}{calendar time to cut the interim/final analysis.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ByTime-validate}{\code{ByTime$validate()}}
\item \href{#method-ByTime-new}{\code{ByTime$new()}}
\item \href{#method-ByTime-summary}{\code{ByTime$summary()}}
\item \href{#method-ByTime-clone}{\code{ByTime$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByTime-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ByTime-validate}{}}}
\subsection{Method \code{validate()}}{
Validate input parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByTime$validate(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{calendar time to cut the interim/final analysis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByTime-new"></a>}}
\if{latex}{\out{\hypertarget{method-ByTime-new}{}}}
\subsection{Method \code{new()}}{
TBA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByTime$new(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{calendar time to cut the interim/final analysis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TBA
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByTime-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ByTime-summary}{}}}
\subsection{Method \code{summary()}}{
TBA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByTime$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ByTime-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ByTime-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ByTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
