% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cut.R
\name{Cut}
\alias{Cut}
\title{Cut class}
\description{
Cut class

Cut class
}
\details{
A class for cuttings in group sequantial designs.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time}}{planned duration cutoff}

\item{\code{event}}{target events to cut the interim/final analysis.}

\item{\code{min_followup}}{minimum follow-up duration}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cut-validate}{\code{Cut$validate()}}
\item \href{#method-Cut-new}{\code{Cut$new()}}
\item \href{#method-Cut-summary}{\code{Cut$summary()}}
\item \href{#method-Cut-clone}{\code{Cut$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cut-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Cut-validate}{}}}
\subsection{Method \code{validate()}}{
Validate input parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cut$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cut-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cut-new}{}}}
\subsection{Method \code{new()}}{
Create a new exponential distribution object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cut$new(time, event, min_followup)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{calendar time to cut the interim/final analysis.}

\item{\code{event}}{target events to cut the interim/final analysis.}

\item{\code{min_followup}}{minimum follow-up duration}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Cut} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cut-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Cut-summary}{}}}
\subsection{Method \code{summary()}}{
Summarize the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cut$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cut-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cut-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cut$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
