% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cut_date.R
\name{get_cut_date}
\alias{get_cut_date}
\title{Simulate stratified time-to-event outcome randomized trial}
\usage{
get_cut_date(y, ...)
}
\arguments{
\item{...}{Additional arguments (not used).}

\item{x}{the object}
}
\value{
TBA
}
\description{
Simulate stratified time-to-event outcome randomized trial
}
