% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdcut-package.R
\docType{package}
\name{gsdcut-package}
\alias{gsdcut}
\alias{gsdcut-package}
\title{gsdcut: Group sequential design with tidy programming grammar.}
\description{
The goal of 'gsdcut' is to offer multiple cutting strategies in group sequential designs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/littlebeannie/gsdcut}
  \item Report bugs at \url{https://github.com/littlebeannie/gsdcut/issues}
}

}
\author{
\strong{Maintainer}: Yujie Zhao \email{yujie.zhao@merck.com}

Other contributors:
\itemize{
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]
}

}
\keyword{internal}
